// fronted/rental-equipment-app-fronted/src/app/services/models/page-response-equipment-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { EquipmentResponse } from '../models/equipment-response';
export interface PageResponseEquipmentResponse {
  content?: Array<EquipmentResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}